% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_geography.R
\name{plot_geography}
\alias{plot_geography}
\title{Plot rates over time by geography.}
\usage{
plot_geography(data)
}
\arguments{
\item{data}{The dataset for the plot}
}
\value{
Plotly object
}
\description{
plot_geography() will create a plotly object.
}
\author{
Braeden Klaver
}
